+++ examples/PerfTest2/PerfTest2.ino
@@ -6,8 +6,8 @@
 #endif
 #if SOC_UART_HP_NUM < 3
   #define Serial2 Serial1
-  #define RX2 RX1
-  #define TX2 TX1
+  #define RX2     RX1
+  #define TX2     TX1
 #endif
 
 // Pin: RX (ESP32) -> TX (JSY)
@@ -56,7 +56,7 @@ void setup() {
     jsy.end();
     jsy.begin(Serial2, RX2, TX2);
 
-    Serial.printf("\njsy.read() at %d bauds:\n", static_cast<int>(rates[i]));
+    Serial.printf("\njsy.read() at %" PRIu32 " bauds:\n", rates[i]);
 
     for (size_t rounds = 1; rounds <= 4; rounds++) {
       Serial.printf(" - ROUND: %d\n", rounds);
